requirements.txt
Flask
Flask-Cors
pandas
google-generativeai
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
